Support scene stylesheets
While Scene Builder can successfully load FXML files containing
javafx.scene.Scene as its root (and by extension, javafx.stage.Stage as
well), the scene's stylesheets do not get applied to the workspace.

This is because we only add the scene's root node to the workspace's
scene graph. As the stylesheet is defined on the scene and not on its
root node, the stylesheet is thus not applied on the workspace.

Fix this by teaching WorkspaceController the concept of
"displayStylesheets", which are stylesheets that should be applied to
the workspace. We then set displayStylesheets to the required values
when loading the Scene so that its stylesheets would be applied to the
workspace.