Background pane: display the fill of the Scene/Window
-fx-background-color needs to be removed from the editor theme CSS
files, otherwise it will override the background pane's background.

To see the effects of this commit, make sure you set the root of the
Scene to a control that does not have its own background (e.g. a Label),
and the Caspian user agent stylesheet is used (as it does not affect the
background fill color of the root node).

See
https://docs.oracle.com/javase/8/javafx/api/javafx/scene/Scene.html#fillProperty
for more info.