LibraryDialogControllerADD LibraryDialog workspace management   - users can link external jars/fxmls or folders to a workspace that is saved in a xml file inside the userlibrary folder.   - ability to import/export a workspace   - command line parsing the workspace: a new switch --ws can be passed to instruct SB to load that workspace (e.g. in Windows launching SceneBuilder.exe --ws="C:\Users\donaldduck\myworkspace.xml")   - the last workspace imported in SB is persistent, meaning that it behaves like regular jars/folders added to the library as before. The workspace can be altered by using an import or by launching SB with the command line switch --ws   - workspace is saved in xml format using XStream library (gradle "com.thoughtworks.xstream:xstream:1.4.11.1")
ADD DocumentWindowController refresh library menuitem: using this command the library is refreshed from scatch; this is especially useful when dealing with changing binaries in linked folders

ADD DocumentWindowController when closing the last non-empty window, a new empty window is opened instead: this avoids closing SB when the last window is closed so that a fresh new window appears. This is behaving more like other softwares (see gimp). If the last empty window is closed, SB terminates as usual.

ADD HierarchyTreeViewController, ContentPanelController pressing ESC on hierarchy treeview or on contentPane selects the parent of the current selected control (like other wysiwyg editors..)

FIX LibraryPanelController the library was trying to copy the same jar file on itself when updating filters after the ImportWindowController dialog

FIX LibraryFolderWatcher the watchService was never stopped and remained active forever; multiple overlapping services (each with its own thread) were started everytime a library update was needed (basically, a watchService.close() was missing in the thread runnable..)