Extract addition of Gluon components
This is the first step towards making the kit module independent of the Gluon charm dependency. The part of BuiltinLibrary's constructor that adds the custom components is changed to acquire them from a component provider instead. The discovery of providers is extracted to a protected method to make the library extendable. In particluar, this enables the Eclipse integration to acquire providers through its extension point mechanism. In BuiltinLibrary, providers are discovered using Java's builtin SPI mechanism.

A new module (com.gluonhq.scenebuilder.components) is added to the project that contains a component provider. The provider is declared in the META-INF/services folder of the components module, in a simple text file. The app now depends on the components module, as well.