The Library Manager ListView is now backed by a SortedList.
A specific comparator class has been created which puts
directories at first, then lists files, then Maven coordinates.

For the sake of testing, a resource (empty file serving as jar)
has been added. With that it can be ensured that the test
works on an existing file and existing folder.

Updated license header.